+++ crates/medea-reactive/src/field/mod.rs
@@ -35,7 +35,7 @@ pub type Observable<D> = ObservableField<D, DefaultSubscribers<D>>;
 /// ([`ObservableField::when`] and [`ObservableField::when_eq`]).
 ///
 /// Can recognise when all updates was processed by subscribers.
-pub type ProgressableField<D> = ObservableField<D, progressable::SubStore<D>>;
+pub type Progressable<D> = ObservableField<D, progressable::SubStore<D>>;
 
 /// Reactive cell which emits all modifications to its subscribers.
 ///
@@ -82,6 +82,8 @@ where
     /// the given `assert_fn` returns `true` on.
     ///
     /// [`Future`]: std::future::Future
+    // TODO: This is kinda broken.
+    //       See https://github.com/instrumentisto/medea/issues/163 issue.
     pub fn when<F>(
         &self,
         assert_fn: F,
@@ -89,12 +91,6 @@ where
     where
         F: Fn(&D) -> bool + 'static,
     {
-        // TODO: This is kinda broken.
-        //       1. Inner value is 0.
-        //       2. let fut = field.when_eq(0).
-        //       3. Change value to 1.
-        //       4. fut.await = Poll::Ready.
-        //       I suggest interior mutability + async fn's.
         if (assert_fn)(&self.data) {
             Box::pin(futures::future::ok(()))
         } else {
@@ -103,7 +99,7 @@ where
     }
 }
 
-impl<D: 'static> ProgressableField<D> {
+impl<D: 'static> Progressable<D> {
     /// Returns new [`ObservableField`] with subscribable mutations.
     ///
     /// Also you can wait for all updates processing by awaiting on
@@ -117,7 +113,7 @@ impl<D: 'static> ProgressableField<D> {
     }
 }
 
-impl<D> ProgressableField<D>
+impl<D> Progressable<D>
 where
     D: Clone + 'static,
 {
@@ -169,6 +165,8 @@ where
     /// value.
     ///
     /// [`Future`]: std::future::Future
+    // TODO: This is kinda broken.
+    //       See https://github.com/instrumentisto/medea/issues/163 issue.
     #[inline]
     pub fn when_eq(
         &self,
@@ -414,7 +412,7 @@ mod tests {
     use futures::{poll, task::Poll, StreamExt as _};
     use tokio::time::timeout;
 
-    use crate::{Observable, ProgressableField};
+    use crate::{Observable, Progressable};
 
     #[tokio::test]
     async fn subscriber_receives_current_data() {
@@ -558,7 +556,7 @@ mod tests {
 
     #[tokio::test]
     async fn when_all_processed_works() {
-        let mut field = ProgressableField::new(1);
+        let mut field = Progressable::new(1);
         assert_eq!(poll!(field.when_all_processed()), Poll::Ready(()));
         *field.borrow_mut() = 2;
         assert_eq!(poll!(field.when_all_processed()), Poll::Ready(()));