+++ crates/medea-reactive/src/subscribers_store/mod.rs
@@ -1,23 +1,23 @@
-//! Implementation of the stores for the updates subscribers.
+//! Stores for updates subscribers.
 
-pub(crate) mod common;
-pub(crate) mod progressable;
+pub mod common;
+pub mod progressable;
 
 use futures::stream::LocalBoxStream;
 
-/// Store for the updates subscribers.
+/// Store for updates subscribers.
 pub trait SubscribersStore<T, O>: Default {
     /// Sends data update to the all subscribers.
     fn send_update(&self, value: T);
 
     /// Creates new updates subscription.
     ///
-    /// Returns [`Stream`] that will yield elements sent with
-    /// [`SubscribersStore::send_update`] calls.
+    /// Returns [`Stream`] yielding elements sent with
+    /// [`SubscribersStore::send_update()`] calls.
     ///
-    /// [`Stream`]: futures::stream::Stream
+    /// [`Stream`]: futures::Stream
     fn subscribe(&self) -> LocalBoxStream<'static, O>;
 
-    /// Wraps provided value to output type.
+    /// Wraps the provided `value` to the output type.
     fn wrap(&self, value: T) -> O;
 }