+++ crates/medea-reactive/CHANGELOG.md
@@ -18,11 +18,11 @@ All user visible changes to this project will be documented in this file. This p
     - `ObservableVec`;
     - `ObservableHashMap`;
     - `ObservableHashSet`.
+- `Progressable` container ([#159]);
 - Progressable collections ([#159]):
     - `ProgressableVec`;
     - `ProgressableHashMap`;
     - `ProgressableHashSet`.
-- `ProgressableField` container ([#159]);
 
 [#81]: /../../pull/81
 [#103]: /../../pull/103